+++ Assets/Scripts/FirstPersonController.cs
@@ -262,7 +262,7 @@ public class FirstPersonController : MonoBehaviour
             ball.TryKick(gameObject);
     }
 
-    private const string GameVersion = "0.3.6-dev";
+    private const string GameVersion = "0.3.7-dev";
 
     private float lastSpeed;
     private bool lastSprinting;